core: rename all Tabs files (with git mv)
