--- packages/core/src/components/Tabs/Tabs.module.css
@@ -1,2 +1,0 @@
-.tabs {
-}
