+++ packages/core/src/components/tabs/index.tsx
